Image Summary in menu with prop update fixes
